+++ src/ServiceNow/ICodeSearchExpandable.ts
@@ -1,0 +1,9 @@
+import { TreeItem } from "vscode";
+
+export interface ICodeSearchExpandable extends TreeItem
+{
+    /**
+     * returns any children of this tree item. 
+     */
+    getChildren(): TreeItem[];
+}