+++ src/ServiceNow/ICodeSearchMatch.ts
@@ -1,0 +1,9 @@
+import { ICodeSearchLineMatch } from "./all";
+
+export interface ICodeSearchMatch
+{
+    field: string;
+    fieldLabel: string;
+    lineMatches: Array<ICodeSearchLineMatch>;
+    count: number;
+}