+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.10.20190802
+* Codesearch is now available through the extension.
+* improve caching on slow network connections. 
+* Added support for:
+  * UI Macroes
+  * Angular Templates
+
 ## 0.9.20190628
 * Fix issue were retrieval of UI Actions would time out causing caching issues.